// src/app/core/enums/tag.enum.ts
export enum Tag {
    // ——— Capacités physiques ———
    Vitesse = 0,
    Endurance = 1,
    Puissance = 2,
    Reflexes = 3,
    Agilite = 4,
    Explosivite = 5,
    Souplesse = 6,
    Coordination = 7,
    Equilibre = 8,

    // ——— Capacités techniques ———
    Technique = 9,
    MaitriseDuBallon = 10,
    Precision = 11,
    Finition = 12,
    PasseDecisive = 13,
    TirCadre = 14,
    Dribble = 15,
    JeuDeTete = 16,
    Tacle = 17,
    Interception = 18,
    GesteTechnique = 19,
    Polyvalence = 20,

    // ——— Intelligence et mental ———
    IntelligenceTactique = 21,
    LectureDuJeu = 22,
    Anticipation = 23,
    Concentration = 24,
    Mental = 25,
    SangFroid = 26,
    Maturite = 27,
    GestionDuTemps = 28,
    Resilience = 29,
    VisionDeJeu = 30,
    Decisif = 31,
    PriseDeDecision = 32,
    Adaptabilite = 33,
    Competiteur = 34,
    Leadership = 35,
    EspritDEquipe = 36,
    Determination = 37,
    ConfianceEnSoi = 38,
    Motivation = 39,
    ApprentissageRapide = 40,
    Discipline = 41,
    Regularite = 42,
    EnduranceMentale = 43,

    // ——— Comportement et personnalité ———
    Travailleur = 44,
    Perseverant = 45,
    Passionne = 46,
    Humilite = 47,
    Exemplaire = 48,
    Respectueux = 49,
    Charismatique = 50,
    Dynamique = 51,
    Engage = 52,
    FairPlay = 53,
    InfluenceurDeVestiaire = 54,
    AttitudeProfessionnelle = 55,

    // ——— Créativité et style ———
    Creativite = 56,
    StyleDeJeuUnique = 57,
    Spectaculaire = 58,
    Showman = 59,
    Imagination = 60,

    // ——— Potentiel / projection ———
    GrosPotentiel = 61,
    EnProgres = 62,
    JeunePrometteur = 63,
    TalentNaturel = 64,
}